add type

+++ entity/userSettings.go
@@ -1,8 +1,8 @@
 package entity
 
 type UserSettingsDb struct {
-	ID          uint `gorm:"primaryKey"`
-	UserID      uint `gorm:"unique"`
+	ID          uint64 `gorm:"primaryKey"`
+	UserID      uint64 `gorm:"unique"`
 	Leagues     uint
 	LevelUser   uint
 	LevelTap    uint
